+++ assets/ThemeAsset.php
@@ -1,0 +1,26 @@
+<?php
+/**
+ * @copyright Copyright (c) 2015 Yiister
+ * @license https://github.com/yiister/yii2-gentelella/blob/master/LICENSE
+ * @link http://gentelella.yiister.ru
+ */
+
+namespace yiister\gentelella\assets;
+
+use yii\web\AssetBundle;
+
+class ThemeAsset extends AssetBundle
+{
+    public $sourcePath = '@vendor/bower/gentelella/build/';
+    public $css = [
+        'css/custom.css',
+    ];
+    public $js = [
+        'js/custom.js',
+    ];
+    public $depends = [
+        'yii\web\YiiAsset',
+        'yii\bootstrap\BootstrapPluginAsset',
+        'rmrevin\yii\fontawesome\AssetBundle',
+    ];
+}